Proposes Finalized Interface Description
In order to continue with the actual work on the OmniView project, there
needs to be a finalized interface, at least for a first release. This
commit should implement this finalization. It doesn't provide
implementation, just an interface-description via a single `.h` file.

The `lib.h` file shows what the `DLL` shall expose in order for a
consumer to utilize a device. For further clarification, take a look at
the [OmniView Product
Vision](https://github.com/skunkforce/OmniView-Product-Vision).

Please review this interface carefully, since it will be here to stay
for a while.

Looking forward to comments!